As per https://lists.freebsd.org/archives/freebsd-scsi/2023-July/000257.html
move to the modern uintXX_t. While here also migrate u_char to uint8_t.
Where other kernel interfaces allow, migrate u_long to uint64_t.

No functional changes intended.

MFC-after:    1 week
Sponsored-by: The FreeBSD Foundation

+++ sys/ufs/ufs/ufs_vfsops.c
@@ -110,7 +110,7 @@ ufs_quotactl(struct mount *mp, int cmds, uid_t id, void * __capability arg,
 			return (EINVAL);
 		}
 	}
-	if ((u_int)type >= MAXQUOTAS)
+	if ((uint64_t)type >= MAXQUOTAS)
 		return (EINVAL);
 
 	switch (cmd) {
